backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
asyncpg
alembic
pydantic>=2.0.0
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
celery[redis]
redis
resend
cryptography
reportlab
supabase
pandas
slowapi
python-dotenv
httpx
orjson
email-validator
